requirements.txt
aws-cdk.core==1.25.0
aws-cdk.aws-ec2==1.25.0
aws-cdk.aws-ecr==1.25.0
aws-cdk.aws-rds==1.25.0
aws-cdk.aws-lambda==1.25.0
aws-cdk.aws-cloudformation==1.25.0
aws-cdk.aws-s3==1.25.0
aws-cdk.aws-iam==1.25.0
aws-cdk.aws-codedeploy==1.25.0
aws-cdk.aws-codebuild==1.25.0
aws-cdk.aws-codecommit==1.25.0
aws-cdk.aws-codepipeline_actions==1.25.0
aws-cdk.aws-datapipeline==1.25.0
aws-cdk.aws-dynamodb==1.25.0
aws-cdk.aws-glue==1.25.0
aws-cdk.aws-secretsmanager==1.25.0
aws-cdk.aws-ssm==1.25.0
aws-cdk.aws-autoscaling==1.25.0
aws-cdk.aws-elasticloadbalancingv2==1.25.0
aws-cdk.aws-servicediscovery==1.25.0
